Rename trial/*.py.
