+++ optuna/trial/__init__.py
@@ -1,5 +1,5 @@
-from optuna.trial.base import BaseTrial  # NOQA
-from optuna.trial.fixed import FixedTrial  # NOQA
-from optuna.trial.frozen import FrozenTrial  # NOQA
-from optuna.trial.state import TrialState  # NOQA
-from optuna.trial.trial import Trial  # NOQA
+from optuna.trial._base import BaseTrial  # NOQA
+from optuna.trial._fixed import FixedTrial  # NOQA
+from optuna.trial._frozen import FrozenTrial  # NOQA
+from optuna.trial._state import TrialState  # NOQA
+from optuna.trial._trial import Trial  # NOQA
