+++ optuna/trial/_state.py
